+++ go.mod
@@ -4,11 +4,11 @@ go 1.22.0
 
 require (
 	github.com/golang-jwt/jwt/v5 v5.2.0
-	github.com/ilyakaznacheev/cleanenv v1.5.0
 	github.com/jmoiron/sqlx v1.3.5
 	github.com/joho/godotenv v1.5.1
 	github.com/lib/pq v1.10.9
 	github.com/redis/go-redis/v9 v9.5.1
+	github.com/sladkoezhkovo/lib v0.1.1
 	golang.org/x/crypto v0.20.0
 	golang.org/x/net v0.21.0
 	google.golang.org/grpc v1.62.0
@@ -20,7 +20,7 @@ require (
 	github.com/cespare/xxhash/v2 v2.2.0 // indirect
 	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
 	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/sladkoezhkovo/lib v0.1.1 // indirect
+	github.com/ilyakaznacheev/cleanenv v1.5.0 // indirect
 	golang.org/x/sys v0.17.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20240123012728-ef4313101c80 // indirect
