+++ internal/entity/role.go
@@ -3,5 +3,5 @@ package entity
 type Role struct {
 	Id        int    `db:"id"`
 	Name      string `db:"name"`
-	Authority uint32 `db:"authority"`
+	Authority int32  `db:"authority"`
 }
